+++ samples/sample-express/src/app.ts
@@ -10,7 +10,7 @@ import {EnableActuator} from "@nodeboot/starter-actuator";
 import {EnableRepositories} from "@nodeboot/starter-persistence";
 import {EnableDI} from "@nodeboot/di";
 import {EnableScheduling} from "@nodeboot/starter-scheduler";
-import {EnableComponentScan} from "@nodeboot/scan";
+import {EnableComponentScan} from "@nodeboot/aot";
 import {EnableHttpClients} from "@nodeboot/starter-http";
 import {EnableValidations} from "@nodeboot/starter-validation";